requirements.txt
# Core Python
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]
asyncpg==0.29.0
greenlet==3.0.3

# Data Processing
polars==0.20.6
pandas
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.0
optuna==3.5.0
shap==0.44.1

# ML Ops
mlflow

# Data Validation
great-expectations==0.18.8

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP
httpx==0.26.0
aiofiles==23.2.1

# Utilities
click==8.1.7
tqdm==4.66.1
joblib==1.3.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0

# Kaggle
kaggle==1.6.6

# Plotting (for notebooks)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
